pyproject.toml
[tool.poetry]
name = "aplicacao"
version = "0.1.0"
description = ""
authors = ["Flavio Carvalho <professorflaviosoares@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-rest-framework = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
